feat: file manager

+++ server/controllers/taskController.js
@@ -3,58 +3,73 @@ const Task = require("../models/task");
 const taskController = {
   addTask: async (req, res) => {
     try {
-      const { id, title, status, label, priority } = req.body;
+      const { title, status, label, priority, assignedTo } = req.body;
+
+      if (!title || !status || !assignedTo) {
+        return res
+          .status(400)
+          .json({
+            message: "Le titre, le statut et l'attribution sont requis.",
+          });
+      }
 
       let newId;
-      if (id) {
-        newId = id; // Use provided ID
-      } else {
-        const lastTask = await Task.findOne({}, {}, { sort: { id: -1 } }); // Get the last task
-        if (lastTask) {
-          const lastId = parseInt(lastTask.id.split("-")[1]);
-          newId = `TASK-${(lastId + 1).toString().padStart(4, "0")}`;
-        } else {
-          newId = "TASK-0001";
-        }
-      }
+      const lastTask = await Task.findOne({}, {}, { sort: { id: -1 } });
+      const lastId = lastTask ? parseInt(lastTask.id.split("-")[1]) : 0;
+      newId = `TASK-${(lastId + 1).toString().padStart(4, "0")}`;
 
       const newTask = new Task({
         id: newId,
         title,
         status,
-        label,
         priority,
+        assignedTo,
       });
       await newTask.save();
-      res.status(200).json({ message: "Tâche ajoutée avec succès." });
+      res.status(201).json({ message: "Tâche ajoutée avec succès." });
     } catch (error) {
       console.error("Erreur lors de l'ajout de la tâche", error);
-      res.status(500).json({ message: "Erreur lors de l'ajout de la tâche." });
+      res
+        .status(500)
+        .json({
+          message: "Erreur lors de l'ajout de la tâche.",
+          error: error.message,
+        });
     }
   },
 
   getAllTasks: async (req, res) => {
     try {
-      const tasks = await Task.find();
+      const tasks = await Task.find().populate("assignedTo"); // Charger les utilisateurs associés
       res.status(200).json(tasks);
     } catch (error) {
       console.error("Erreur lors de la récupération des tâches", error);
       res
         .status(500)
-        .json({ message: "Erreur lors de la récupération des tâches." });
+        .json({
+          message: "Erreur lors de la récupération des tâches.",
+          error: error.message,
+        });
     }
   },
 
   removeTask: async (req, res) => {
     try {
       const taskId = req.params.id;
+      const task = await Task.findOne({ id: taskId });
+      if (!task) {
+        return res.status(404).json({ message: "Tâche non trouvée." });
+      }
       await Task.deleteOne({ id: taskId });
-      res.status(200).json({ message: "Tâche supprimée avec succès." });
+      res.status(204).json({ message: "Tâche supprimée avec succès." });
     } catch (error) {
       console.error("Erreur lors de la suppression de la tâche", error);
       res
         .status(500)
-        .json({ message: "Erreur lors de la suppression de la tâche." });
+        .json({
+          message: "Erreur lors de la suppression de la tâche.",
+          error: error.message,
+        });
     }
   },
 };
